Modeling default coef probs with distribution
Replaces the default tables for single coefficient magnitudes with
those obtained from an appropriate distribution. The EOB node
is left unchanged. The model is represented as a 256-size
codebook where the index corresponds to the probability of the
zero node on the coef encoding tree. The node probabilities
are then obtained by looking up the codebook using the prob of
the zero node as the index. The main advantage is that the
default prob tables will become considerably smaller and manageable,
Besides there is substantially less risk of over-fitting for a
training set.

Various distributions are tried and the one that gives the best
results is the family of Generalized Gaussian distributions with
shape parameter 0.75. The shape parameter for Y && INTRA planes may
be chosen to be smaller than the other components.
The results are within about 0.2% of fully trained tables.

The forward updates are optionally (controlled by a macro)
model-based, i.e. restricted to only convey probabilities from the
codebook. Backward updates can also be optionally (controlled by
another macro) model-based, but is turned off by default. Currently
model-based forward updates work about the same as unconstrained
updates, but there is a drop in performance with backward-updates being
model based.

The model based approach also allows the probabilities for the key frames
to be adjusted from the defaults based on the base_qindex of the
frame. Currently the adjustment function is a placeholder that adjusts
the prob of EOB and 0 node from the nominal one at higher quality
(lower qindex) or lower quality (higher qindex) ends of the range. The
rest of the probabilities are then derived based on the model from the
adjusted prob of zero.

Change-Id: Iae050f3cbcc6d8b3f204e8dc395ae47b3b2192c9