Models coef probs with a suitable distribution
Replaces the tables for single coefficient magnitudes with
those obtained from an appropriate distribution. The EOB node
is left unchanged. The model is represented as a 256-size
codebook where the index corresponds to the probability of the
zero node on the coef encoding tree. The node probabilities
are then obtained by looking up the codebook using the prob of
the zero node as the index.

Various distributions are tried and the one that gives the best
results is the Generalized Gaussian with shape parameter 0.75.
The results are quite close to trained tables with the max
difference being -0.265% for hd, and yt even improving by +0.206%.

Generalized Gaussian 0.75:
derf: -0.080%
yt: +0.206%
stdhd: -0.133%
hd: -0.265%

TODO - forward/backward updates, optimize the codebook of
coefficient probs.

Change-Id: Iae050f3cbcc6d8b3f204e8dc395ae47b3b2192c9