Models coef probs with a suitable distribution
Replaces the tables for single coefficient magnitudes with
those obtained from an appropriate distribution. The EOB node
is left unchanged. The model is represented as a 256-size
codebook where the index corresponds to the probability of the
zero node on the coef encoding tree.

WIP.
TODO - forward/backward updates, optimize the codebook of
coefficient probs - may need training as opposed to modeling.

Change-Id: Iae050f3cbcc6d8b3f204e8dc395ae47b3b2192c9