Update patch set 4
Patch Set 4:

I'm trying to understand the coefficient optimization changes you made, I'm not quire sure I understand them yet, I'll try to pick your mind on monday regarding that. I think it only tests context-costing changes on rounding changes if we are a context for tokens directly following us, whereas the effect of this change is that we can be a context for tokens much further down the line. I think that may well mean a semi-rewrite for optimize_b, but like I said, I don't fully understand that code.

Patch-set: 4