Zero group experiment
Starts an experiment that codes a group of zeros in a block
following one of several patterns. Once a zero is transmitted an
additional  zerogroup symbol will be sent to indicate if it is
an isolated zero, or a zerotree root, or end of orientation
(vert/hor/diag) i.e. all coeffs in the same orientation are zero
starting from the current one in the block. The coefficients
indicated as zero as part of zerogropup information will be
skipped when they are encountered in the scan order.

Later we can even fold in the EOB symbol within this zerogroup
symbol, but holding off on that to keep the experiment manageable
in terms of code changes.

WIP: very prelimiary. Feedback on the approach appreciated.

Change-Id: I7974ac40806eda2ae899aa51a75e45e1ad98bf9c