Zero group experiment
Starts an experiment that codes a group of zeros in a block
following one of several patterns. Once a zero is transmitted
additional zerogroup symbols will be sent to indicate if it is
an isolated zero, or a zerotree root, or end of orientation -
i.e. all coeffs in the same orientation (vert/hor/diag) are zero
starting from the current coefficient in the block. The coefficients
indicated as zero as part of a prior zerogropup information will be
skipped when they are encountered in the scan order.

Later we can even fold in the EOB symbol within this zerogroup
symbol, or even remove the EOB symbol altogether since sending
a EOB symbol may be just as efficient as sending three End of
Orientation symbols successively. But I will be holding off on those
variations to keep the experiment manageable in terms of code changes.

WIP:
STATUS: Encode/decode sync is obtained.
TODO: prob updates, training for default probs.

Change-Id: I7974ac40806eda2ae899aa51a75e45e1ad98bf9c