Zero group experiment
Starts an experiment that codes a group of zeros in a block
following one of several patterns. Once a zero is transmitted
additional zerogroup symbols will be sent to indicate if it is
(1) an isolated zero, or (2) end of orientation - i.e. all coeffs
in the same orientation (vert/hor/diag) are zero starting from the
current coefficient in the block, or (3) a special zero pattern.
The zero pattern could be either a zerotree root or a run of
8 zeros. The coefficients indicated as zero as part of prior
zerogropup information will be skipped when they are encountered
in the scan order.

The patch is designed to include a variety of configurable
options - whether zerotree or zeroruns are used, what the length
of the zerorun should be, and which extra symbols are used for
which transform sizes.

Later we can even fold in the EOB symbol within this zerogroup
symbol, or even remove the EOB symbol altogether since sending
a EOB symbol may be just as efficient as sending three End of
Orientation symbols successively. But I will be holding off on those
variations to keep the experiment manageable in terms of code changes.

Results (with end-of-orientation for transforms 16x16 and higher,
and an additional zerotree root symbol for 32x32 transforms) are
within 0.15% on all four sets. The probabilities have not been
trained yet.

derf: -0.061%
yt: -0.041%
hd: -0.137%
stdhd: -0.089%

Results with the zerotree root symbol replaced by a run of n
zeros, is slightly worse.

Experiments with various other configurable options and various
variations are in progress.

Change-Id: I7974ac40806eda2ae899aa51a75e45e1ad98bf9c