Initial addition of multiple ARF frames
*** This is work-in-progress, not ready for review ***

It adds the ability to
insert multiple ARF frames into a single ARF group. This
patch implements the reordering of the coded frames, and
implements a fixed-length coding pattern. It applies a
fixed quantizer strategy based on where the frame is in
the coding sequence.

Further work to modify the rate control strategy is
ongoing and will be submitted via a set of future patches.

In this first step, each ARF group is recursively
bisected and an ARF frame added at that position in the
sequence. The recursion continues until ARF frames are
within MIN_GF_INTERVAL frames.

The code sits behind the "multiple-arf" experimental
flag ("CONFIG_MULTIPLE_ARF").

Change-Id: Ie473b05ebb43ac473c0cfb659b2b8042823085e2