Update patch set 1
Patch Set 1:

This one actually breaks the newbintramodes experiment since the context pred mode requires that we have reconstructed the 4x4 blocks on the left and top. That was the reason for special casing the B_PRED mode.

At this point, I would suggest one of two things: either revert this change, or remove the NEWBINTRAMODES experiment (since it gives only small gains).

Patch-set: 1