Adaptive scantables.
Fairly rough, allows decoding 3 coefficients and adapts scantable if
the energy in the horizontal/vertical ACs suggests a bias in either
one direction.

Trellis is disabled when this experiment is on, because trellis could
change the scantable. I may be able to re-enable it but skip trellis
over the first 3 coefficients or something along those lines, have to
do some tests.

HD gains 0.2%. Other sets are mostly unchanged.

Change-Id: Iff2d0f58e856134e555d2e0e3cbfd0f00f717d39