Update patch set 1
Patch Set 1:

Found this mv clamping issue when testing on performance checkpoint for rectangular superblock coding modes. It affects the prediction efficacy of NEWMV mode for those blocks sitting close to the frame boundary, and hence the coding performance. It is resolved in this patch.

Patch-set: 1
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0