Update patch set 3
Patch Set 3:

(Not objecting to the patch), I wonder why the clamp would change the motion values for e.g. NEWMV if the input value is guaranteed to fit within the clipping range? Shouldn't that effectively make it a no-op? So why does removing the clip then change behaviour at all?

To me, this suggests that perhaps the clipping range is actually incorrect, which may even be a bitstream bug (assuming the decoder and encoder both agree on that being the correct clipping range).

Patch-set: 3
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0