Update patch set 3
Patch Set 3: I would prefer that you didn't submit this

So on a more general note, we need to audit every location where we do clamping. The clamping rules in VP8 are quite perverse (read: buggy) and I'd like to avoid repeating that.

Removing this clamp on the decoder doesn't look right to me. Does this work if you remove it from the encoder only?

Take a look at this commit for a little history on what need_to_clamp_mv was about in VP8:

commit 3085025fa1392e99bc95a519657374f2e9a4249b
Author: John Koleszar <jkoleszar@google.com>
Date:   Wed Jun 9 11:29:20 2010 -0400

    Remove secondary mv clamping from decode stage

It was only a security feature in VP8 (which is why I don't want it removed from the decoder hastily) that said if you have an MV that points way into the border, it gets clamped to a point we can actually read. The VP8 encoder doesn't rely on it in any way. The MV used for entropy and relative coding purposes was left unclamped (beyond what the bitstream required), and could point arbitrarily far into the border. need_to_clamp was set if it pointed so that the whole block was in the UMV border, so that we could equivalently predict from a shorter vector. The dixie decoder doesn't need it, because it uses emulated edges rather than constructing a border, but they're equivalent -- the UMV border can be considered infinite.

In VP9, it's not necessarily a bug if we're relying on it -- for a 64x64 SB, you could be reading more from the UMV than the 64 pixels we populate. It might be interesting to try changing the border constant in VP9 back to 32 and test the decoder decodes equivalently. Things in theory should work just as they do now since I think we only ever filter on a 16x16, and this need_to_clamp behavior should handle the larger SBs, if we're using it right. Might find some bugs that way. Can't do it on the encoder though, since we use it for motion search.

Patch-set: 3
Label: Code-Review=-1
Label: Commit-Queue=0
Label: Verified=0