Update patch set 5
Patch Set 5:

The confusion is that I think that this refactor will hurt get Jingning's non-square block size support in, and in general will make it more difficult to support non-square blocks. It makes a lot of assumptions that are not true in non-square blocks.

Is it possible to leave some of these major refactors for a little bit until the block coding structure that Jingning and I are working on has crystalized a little bit?

Patch-set: 5
Reviewer: Gerrit User 1000885 <1000885@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0