Update patch set 5
Patch Set 5:

Citation needed. I don't think that it makes it hard to support non-square blocks -- that was always intended to be supported (and is already, for 4:2:2). It's probably only a couple lines to make it work for non-square blocks if it doesn't already. It only really assumes square transforms. In any case, the non-square form needs to be expressible concisely like this, so maybe this is a good starting point.

This is ready to go now, the other change isn't. This one is small, and basically contained to one file. The other change doesn't appear to touch this file, so I don't see what the conflict is at all yet. If it does, you can easily wrap it with if (!square) if you need to. I don't think we should hold it back for another change that for the most part it doesn't conflict with. I don't want to build up a whole string of changes that will be impossible to merge. I appreciate that you're faced with that right now, and I'm not suggesting that one is more important than the other. They're both equally important, and I could argue on the opposite side about the assumptions that patch makes about U&V. This is why you keep changes small.

Patch-set: 5
Reviewer: Gerrit User 1000856 <1000856@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0
Label: Verified=0