Remove the mb_no_coeff_skip flag
This flag was added to VP8 to allow a mode where MB-level skipping
was not allowed, saving a bit per mb. It was never used in practice,
and hasn't been tested in VP9, so remove it.

This does change the bitstream slightly, since the top-level bit
is removed, so it's probably best to wait for another change of
this type to submit it. It'll need a minor rebase after mb and
sb functions are combined, but posting for discussion for now.

Change-Id: Id450ec6904c6d06c1919508e7efc52d05cde5631