Merge loop over all macroblock modes into encode_sb_row().
Rename pick_mb_modes to pick_mb_mode, since it now handles only a
single macroblock. This is consistent with pick_sb_mode handling a
single non-macroblock.

Quality improves slightly because we used to not initialize the
storage context for the temp encode of the macroblock while RD
searching subsequent macroblocks.

Change-Id: I896fdfa06436b2d8c24d6474718cc74420df6b3b