Update patch set 2
Patch Set 2:

As for function names, can we keep the sby/uv prefix for consistency with other functions, to indicate it acts on a arbitrary-sized recursive block of data (16x16, 32x32, 64x64, perhaps in the future smaller or rectangular)? Removing the tx_size as an argument is fine with me but you'll have to fix the cases where mbmi->txfm_size isn't set, and ideally not using the hack you introduced in quantize (it should be possible to remove that since I merged the various super/macro_block_yrd functions).

I don't particularly like the fact that we force arguments on the stack (there's good reason x86-64 does that in registers instead), but that's something we can fix later, it's more a performance thing than anything functional (and minor anyway).

Patch-set: 2