Update patch set 3
Patch Set 3:

I agree that there's some perf concern as it defeats inlining, but also that it can be fixed later pretty easily. At this point, the goal is to minimize the code and make it provably consistent and general where it can be.

Patch-set: 3
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0