Update patch set 1
Patch Set 1:

I think you may have misunderstood the scope of what I was suggesting, and perhaps I wasn't clear when I said asm level before, as I was really talking about rtcd. Keeping the various block sizes in the asm (not rtcd) level is ok for now. What I'm suggesting right now is keep everything we're doing the same, just put behind a common function at the rtcd level. Just move the function pointer table from cpi into the rtcd layer. So rather than:

  block_size = map_bsize_to_block_size_enum_y(bsize)
  cpi->fn_ptr[block_size].sad(...)

you write:

  vp9_sad_sby(..., bsize);

and vp9_sad_sby_c() becomes:

  void vp9_sad_sby_c(..., bsize) {
    int w = bsize_to_w(bsize), h = bsize_to_h(bsize);
    sad_mxn(w,h,...)
  }

and vp9_sad_sby_sse2() becomes:

  void vp9_sad_sby_sse2(..., bsize) {
    const fn_ptr[] = {sad16x16_sse2, sad32x16_sse2, sad32x32_sse2, ...}
    idx = map_bsize_to_idx(bsize)
    fn_ptr[idx](...)
  }

or do this with a switch(bsize) to avoid the map_bsize_to_idx(), which can go away after the bsize enum is final.

It should be up to the rtcd/asm layer to decide how the functions are implemented -- either a byte at a time, or one function per register width with h as a parameter, or one function per block size, or maybe something else entirely.

We can fix this later, but since you're in this code now and it's as much or less work to fix the layering violation as it is to cut and paste all this code, I wanted to point it out. In particular, if you find yourself needing to do this for another block size dependent rtcd function, keep this in mind.

Patch-set: 1