Masked compound-inter experiment
This is the first of a series of experiments that uses masked motion
in compound inter mode.

In this expt, an index is sent from a mask codebook to convey a
mask when two motion vectors are used for blocks 16x16 and larger. The
mask is used to weight pixels from the first and second predictors to
obtain the final predictor. The weighting is smooth near the partition
boundaries but becomes a selecton farther away. This first implementation
does not use a masked motion search, but instead just combines the best
first and second mvs obtained independently to produce masked predictors
for different masks.

The current mask codebook has mask types separating a block into wedges
at specific angles and offsets. For 16x16 macroblocks, an 8-ary mask
codebook is used, while for larger superblocks a 16-ary codebook is
used. Results indicate that at larger blocks finer granularity of
directions may be needed to have more benefits.

Masked seach code is in place, but a full masked search is very
expensive for the encoder. Instead a refinement masked search is used if
the masked search is found to be better than standard compound inter
coding in the rd loop.

Results:
derf300: +0.56%

Change-Id: Ia78a258ba9d7364dc43933ee5a95a3f9de43e1c1