Masked compound-inter experiment
This is the first of a series of experiments that use masked motion
in compound inter mode.

In this expt, an index is sent from a mask codebook to convey a
mask when two motion vectors are used for blocks 16x16 and larger. The
mask is used to select pixels from the first and second predictors to
obtain the final predictor. This first implementation does not use a
masked motion search, but instead just combines the best first and
second mvs obtained independently to produce masked predictors for
different masks.
The current mask codebook has 8 mask types separating a block
into wedges passing through the center and at slopes 1/2, -1/2, 2
and -2 respectively, and their inverses.

Results: pending

TODO:
Masked search (will increase encode time significantly, but should
improve performance).
Filtering of the combined predictor across discontinuities in the mask.
Replace mask codebook with other means of conveying mask including
sending an additional MV, seed color, etc.

Change-Id: Ia78a258ba9d7364dc43933ee5a95a3f9de43e1c1