Masked compound-inter experiment
In this expt, an index is sent from a mask codebook to convey a
mask when two motion vectors are used for superblocks. The
mask is used to weight pixels from the first and second predictors to
obtain the final predictor. The weighting is smooth near the partition
boundaries but becomes a selecton farther away.

The current mask codebook has mask types separating a block into wedges
at specific angles and offsets. The number of bits used for the mask
codebook can be 3, 4, or 5 depending on the size of the superblock.
Results indicate that at larger blocks finer granularity of directions may be
needed to have more benefits.

Masked search code is in place, but a full masked search is very
expensive for the encoder. Instead a refinement masked search is used if
the masked search is found to be better than standard compound inter
coding in the rd loop.

Results on May 22 head:
derfraw300: +1.030%
stdhdraw250: pending

Change-Id: Ia78a258ba9d7364dc43933ee5a95a3f9de43e1c1