Masked compound-inter 3mv experiment
This is the second of a series of experiments that uses masked motion
in compound inter mode. In this variation, 3 motion vectors are used,
where two of them are associated with the second ref-frame. The second
MV (associated with the second ref-frame) is used to decide on a mask,
while the third MV (also associated with the second ref-frame) is used
for actual prediction of the masked pixels.
The final predictor is comprised by the unmasked pixels predicted from
the first MV and first ref-frame and masked pixels predicted from the
third MV and second ref-frame.

Status: Full encode-decode sync achieved for the first pass.
Results: pending.

Change-Id: I68742f1320dc99a1e010ab99b0a6cd8ea9435d0e