Some more minor loopfilter changes for rectangles and variable blocksizes.
Adjust txfm_size to be updated every time we change mbmi. I don't
think this has any practical effect (since we check that it's >32x32,
which means it was the same for the whole SB), but it's more consistent.
Then, don't use sb_mb_lf_skip, because it's only called for tx>=32x32,
and in those cases all conditions are automatically true and the only
thing we need to check is the skip flag (mb_skip_coeff). Lastly, support
64x32 and 32x64 rectangles also.

Change-Id: Ie3fbdccb3da256a101e1a965579404d69c2063f2