Update patch set 4
Patch Set 4:

This one appears to give identical output before/after and with/without patch. Once this is in, above/left_context need to be converted to use an 8x8 basis (I have a partial patch for that somewhere), we need to add a 8x8/8x16/16x8 iterations to encodeframe RD loop, enable bitstream writing/reading and reconstruction of such blocksizes.

Patch-set: 4
Label: Commit-Queue=0
Label: Verified=0