Update patch set 2
Patch Set 2:

There are a few things to think about here, depending on how we plan to implement the smaller block sizes. If we're going to introduce a 2x2 transform, then this is probably ok, as long as the intention of adding a bsize argument is to make a common function that works for all block sizes. Otherwise, this should be a new 2x2 function, consistent with 1 function per transform size like the other code. 

If we're going to stick with 4x4 as the smallest transform size, then this change is probably unnecessary, as idct-add should map to the transform size, not the prediction size. It seems like some/all of these functions should go away, as we move towards the existing common superblock code. The 4x4 path should be coerced through xd->itxm_add and vp9_iht_add_c.

I've got a related patch that unifies the idct-add stuff in decode frame for all the block sizes. It's not quite ready for formal review, but it's worth taking a look at in the context of this change. I'll post so you can take a look.

Patch-set: 2
Reviewer: Gerrit User 1000856 <1000856@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0