Update patch set 1
Patch Set 1: (1 inline comment)

Shouldn't be part of this commit, but I think we probably want to make tx_type a per-plane element, and then we can use your function that calculates the uv tx size to initialize it for each plane. That would move the is_split stuff out of the foreach_ walkers, and be more readable.

I still think that doing 8x8 chroma transform given a 4x4 luma prediction is sometimes a bad choice. If we can bake that logic into one place, then it'd be easy to try both ways later. I also may be looking at a more generic chroma tx size selection as part of 4:4:4, to allow you to either match the tx size to the subsampling or try to maximize it, as we do now in some modes. For 4:4:4, I'll likely start by using the same size as the luma, but there are probably cases for using a larger chroma than luma transform (just like we do now) so centralizing this logic would be a win.

Patch-set: 1
Label: Code-Review=0
Label: Commit-Queue=0