Balancing coef-tree to reduce bool decodes
This is an initial patch that changes the coefficient tree to move
the EOB to below the ZERO node in order to save bool decodes.

The advantages of moving EOB one step down as opposed to two steps down
in the other parallel patch are that: 1. The coef modeling based on
the One-node becomes independent of the tree structure above it, and
2. Fewer conext/counter increases.
The drawback is that the potential savings in bool decodes will be
less, but assuming that 0s are much more predominant than 1's the
potential savings is likely to be less.

TODO: Add separate counts and contexts for the zero node depending
on whether the eob node is skipped or not.

Change-Id: Ie784be13dc98291306b338e8228703a4c2ea2242