Update the neighbors searched for MV condidates
The ultimate goal of this patch is to find a way to remove the MV from
the last frame as a candidate to allow that dependency to be
eliminated. Since just removing that MV causes a nontrivial quality
loss, that MV is being used often. Given that there are 8 spatial
neighbors considered before using the last frame's MV and only two
candidates allowed, it seems that there is a lot of overlap in the
candidate MVs, if two distinct candidates aren't being found. This
patch expands the search range such that the candidates are more likely
to be in different macroblocks.

Change-Id: I7302b5bb38c7a86f41fc705c8d193e5bc9a40dc8