Implict Segmetation further small patch.
Some renaming but still for now under config flag.

Set flag to allow or disallow in speed features function.

Off by default for now.

Tested with feature on and flag on and off. The results
in both cases were neutral as expected because the Q
adjustment is currently very small.

This feature is as yet of unproven worth but may prove valuable
in the future when we try adjusting other features like loop filter or
looking at psychometric rather than metric impact.

Tested to

Change-Id: Icea62d568f58e7fb55fc5d7540bff0a96577d072