Update patch set 1
Patch Set 1:

The title of the patch is a little weird, it suggests you're fixing an actual buffer overflow, which confused me, I guess. Maybe change the commit msg to reflect what the patch does (it makes bmi smaller, but does so without introducing a buffer overflow, because it makes the firstpass - which uses 16 4x4 subblocks with a 16x16 primary structure - no longer use bmi[] at all).

(Or something like that.)

Patch-set: 1