Reimplementation of loop filter
This version of the loop filter supports non-4:2:0 subsampling and
a fourth plane, as well as changing the filtering order to be more
friendly to SIMD and hardware implementations.

The filters are applied first to all vertical edges within the
64x8 mi-row, followed by the top horizontal edge and any internal
horizontal edges. This application order for the vertical filters
is changed to remove some serial dependencies present in the
previous implementation. There is still a dependency when doing
the 4x4 filtering. It may be possible to remove that.

It may be more desirable to apply all vertical edges first in the
64x64 block, but it requires an additional pass through the data
or additional storage for the horizontal edge masks. Want to get
feedback on this implementation first.

Results are slightly positive on derf (+.012), other sets pending.

Change-Id: I08f3e9683eb7bded8a76651cbc50fc0dfdd05fa7