Change order in rd_pick_partition()
Order of operations changed in recursive loop such that
we abort the recursive split down if the PARTITION_NONE
case is as good (both in rd and distortion) as either the
vertical or horizontal partition cases.

This behavior is controlled by a speed feature flag
cpi->sf.allow_partition_split and is currently for test purposes
only enabled at good quality speed 1 and above.

The speed impact is likely to be modest on derf (cif) where most
blocks will need to split down quite small, but will likely be
more substantial for image larger formats.

Results pending for quality impact.

Change-Id: Id79ddda33c70f3b9693516819376a994ccee6da3