Call wider loopfilters where possible
Do 16 pixels at a time if both 8x8 blocks use the same filter level.
Mostly for discussion at this point, and example of the necessity
for checking that the lfi's are the same for both blocks. This seems
slightly slower for the current code, so long as the filters are
still in C.

Change-Id: Ic3643cc7c8e6dcc2b1f58d1fb47a42ac57e2d296